# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Rest RCS Express",
    "description": "API Rest RCS Express",
    "contact": {
      "name": "rcsxpress",
      "email": "operaciones@rcsexpress.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Ambiente de Desarrollo"
    },
    {
      "url": "https://scen.rcsexpress.com/api/v1/",
      "description": "Ambiente de Producción"
    }
  ],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/cuentas": {
      "get": {
        "tags": ["Cuentas"],
        "summary": "Obtener todos las cuentas",
        "description": "API para Obtener todos las cuentas del sistema",
        "parameters": [
          {
            "name": "banco",
            "in": "header",
            "description": "Codigo del Banco",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cuentas"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Cuentas"],
        "summary": "Crear Cuentas",
        "description": "API para Crear una nueva Cuenta",
        "requestBody": {
          "description": "Post object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cuentas"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cuentas"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      }
    },
    "/cuentas/{id}": {
      "get": {
        "tags": ["Cuentas"],
        "summary": "Obtener Cuenta por su Id",
        "description": "API para Retornar una Cuenta por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo de la Cuenta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cuentas"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Cuentas"],
        "summary": "Actualizar Cuentas",
        "description": "API para Actualizar una Cuenta por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo de la Cuenta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Put object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cuentas"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cuentas"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Cuentas"],
        "summary": "Borrar Cuentas",
        "description": "API para Borrar una Cuenta por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo de la Cuenta",
            "required": true, "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/Cuentas/properties/id"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      }
    },
...
    "/estados": {
      "get": {
        "tags": ["Estados"],
        "summary": "Obtener todos los estados",
        "description": "API para Obtener todos los estados del sistema",
        "parameters": [
          {
            "name": "pais",
            "in": "header",
            "description": "Codigo del Pais",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estados"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Estados"],
        "summary": "Crear Estados",
        "description": "API para Crear un nuevo Estado",
        "requestBody": {
          "description": "Post object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Estados"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estados"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      }
    },
    "/estados/{id}": {
      "get": {
        "tags": ["Estados"],
        "summary": "Obtener Estado por su Id",
        "description": "API para Retornar un Estado por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Estado",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estados"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Estados"],
        "summary": "Actualizar Estado",
        "description": "API para Actualizar un Estado por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Estado",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Put object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Estados"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estados"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Estados"],
        "summary": "Borrar Estado",
        "description": "API para Borrar un Estado por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Estado",
            "required": true, "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/Estados/properties/id"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cuentas": {
        "description": "Modelo de Cuentas",
        "required": ["id", "nro_cuenta", "cod_banco"],
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "nro_cuenta": {
            "type": "string"
          },
          "tipo_cuenta": {
            "type": "string"
          },
          "firma_autorizada": {
            "type": "string"
          },
          "flag_activa": {
            "type": "string"
          },
          "cod_banco": {
            "type": "number"
          },
          "tipo_desc": {
            "type": "string",
            "readOnly": true
          },
          "activa_desc": {
            "type": "string",
            "readOnly": true
          }
        }
      },
...
      "Estados": {
        "description": "Modelo de Estados",
        "required": ["id", "desc_estado", "siglas", "iso_3166", "cod_pais"],
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "desc_estado": {
            "type": "string"
          },
          "siglas": {
            "type": "string"
          },
          "iso_3166": {
            "type": "string"
          },
          "cod_pais": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}